requirements.txt
coreapi
django
django-rest-swagger
djangorestframework
dynamic-rest
Pillow